requirements.txt
asgiref==3.5.2
certifi==2022.9.14
distlib==0.3.4
Django==3.2.15
django-dotenv==1.4.2
filelock==3.7.1
pipenv==2022.9.8
platformdirs==2.5.2
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
virtualenv==20.14.1
virtualenv-clone==0.5.7
psycopg2-binary
gunicorn
django-k8s